[FEAT] # 20 - 게시글 클러스터링 이후 마커 클릭시 마커에 연동된 게시물 확인
Former-commit-id: e0b54cc978a0d26b2cf1a0c515d16a7d6fc368cf

+++ Smile_Server/api-module/src/main/java/com/ssafy/api/dto/article/ArticleClusterDto.java
@@ -4,11 +4,14 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 
+/***
+ * 클러스터링시 포함된 게시글의 개수, 중심좌표 반환
+ */
 @Builder
 @Data
 @AllArgsConstructor
 public class ArticleClusterDto {
-    private Integer clusterId;
+    private Long clusterId;
 
     private Integer numOfCluster;
